# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BotSource = typeof BotSource[keyof typeof BotSource];


export const BotSource = {
  github: 'github',
  zip: 'zip',
} as const;

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  online: 'online',
  offline: 'offline',
  deploying: 'deploying',
} as const;

export interface Bot {
  id: string;
  name: string;
  title: string;
  description: string;
  source: BotSource;
  sourceLabel?: string;
  status: BotStatus;
  memoryMb: number;
  updatedAt: string;
  /** @nullable */
  lastDeployedAt: string | null;
  uptime?: string;
  monthlyUsage?: number;
}

export type BotInputSource = typeof BotInputSource[keyof typeof BotInputSource];


export const BotInputSource = {
  github: 'github',
  zip: 'zip',
} as const;

export interface BotInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  title: string;
  description: string;
  source: BotInputSource;
  sourceLabel?: string;
  /** @minimum 1 */
  memoryMb?: number;
}

export interface BotUpdate {
  name?: string;
  title?: string;
  description?: string;
  /** @minimum 1 */
  memoryMb?: number;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  deploy: 'deploy',
  status: 'status',
  edit: 'edit',
  upload: 'upload',
} as const;

export interface Activity {
  id: string;
  type: ActivityType;
  title: string;
  detail: string;
  timestamp: string;
}

export interface Dashboard {
  onlineBots: number;
  totalBots: number;
  maxBots: number;
  memoryUsedMb: number;
  memoryLimitMb: number;
  monthlyHours: number;
  planName: string;
  uptimePercent: number;
}

export interface CheckoutInput {
  email: string;
}

export interface CheckoutSession {
  id: string;
  url: string;
}