# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard overview operations
  - name: bots
    description: Bot hosting operations
  - name: activity
    description: Recent activity operations
  - name: billing
    description: Billing and checkout operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard overview
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /bots:
    get:
      operationId: getBots
      tags: [bots]
      summary: List hosted bots
      responses:
        "200":
          description: Hosted bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Create a hosted bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Created bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
  /bots/{id}:
    patch:
      operationId: updateBot
      tags: [bots]
      summary: Update bot metadata
      parameters:
        - $ref: "#/components/parameters/BotId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotUpdate"
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Bot not found
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a hosted bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "204":
          description: Deleted bot
        "404":
          description: Bot not found
  /bots/{id}/toggle:
    post:
      operationId: toggleBot
      tags: [bots]
      summary: Toggle a bot online or offline
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          description: Updated bot status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Bot not found
  /activity:
    get:
      operationId: getActivity
      tags: [activity]
      summary: Get recent activity
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /billing/pro-checkout:
    post:
      operationId: createProCheckout
      tags: [billing]
      summary: Create a hosted Stripe checkout for Hollow Cloud Pro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Hosted checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "400":
          description: Invalid checkout details
        "502":
          description: Payment provider unavailable
components:
  parameters:
    BotId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Bot:
      type: object
      required: [id, name, title, description, source, status, memoryMb, updatedAt, lastDeployedAt]
      properties:
        id:
          type: string
        name:
          type: string
        title:
          type: string
        description:
          type: string
        source:
          type: string
          enum: [github, zip]
        sourceLabel:
          type: string
        status:
          type: string
          enum: [online, offline, deploying]
        memoryMb:
          type: integer
        updatedAt:
          type: string
        lastDeployedAt:
          type: ["string", "null"]
        uptime:
          type: string
        monthlyUsage:
          type: integer
    BotInput:
      type: object
      required: [name, title, description, source]
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        description:
          type: string
        source:
          type: string
          enum: [github, zip]
        sourceLabel:
          type: string
        memoryMb:
          type: integer
          minimum: 1
    BotUpdate:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        description:
          type: string
        memoryMb:
          type: integer
          minimum: 1
    Activity:
      type: object
      required: [id, type, title, detail, timestamp]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [deploy, status, edit, upload]
        title:
          type: string
        detail:
          type: string
        timestamp:
          type: string
    Dashboard:
      type: object
      required: [onlineBots, totalBots, maxBots, memoryUsedMb, memoryLimitMb, monthlyHours, planName, uptimePercent]
      properties:
        onlineBots:
          type: integer
        totalBots:
          type: integer
        maxBots:
          type: integer
        memoryUsedMb:
          type: integer
        memoryLimitMb:
          type: integer
        monthlyHours:
          type: integer
        planName:
          type: string
        uptimePercent:
          type: number
    CheckoutInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    CheckoutSession:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
        url:
          type: string