# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from './activityType';

export interface Activity {
  id: string;
  type: ActivityType;
  title: string;
  detail: string;
  timestamp: string;
}