# lib/api-zod/src/generated/types/bot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotSource } from './botSource';
import type { BotStatus } from './botStatus';

export interface Bot {
  id: string;
  name: string;
  title: string;
  description: string;
  source: BotSource;
  sourceLabel?: string;
  status: BotStatus;
  memoryMb: number;
  updatedAt: string;
  /** @nullable */
  lastDeployedAt: string | null;
  uptime?: string;
  monthlyUsage?: number;
}