# lib/api-zod/src/generated/types/botInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotInputSource } from './botInputSource';

export interface BotInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  title: string;
  description: string;
  source: BotInputSource;
  sourceLabel?: string;
  /** @minimum 1 */
  memoryMb?: number;
}