# lib/api-zod/src/generated/types/botUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotUpdate {
  name?: string;
  title?: string;
  description?: string;
  /** @minimum 1 */
  memoryMb?: number;
}