# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Dashboard {
  onlineBots: number;
  totalBots: number;
  maxBots: number;
  memoryUsedMb: number;
  memoryLimitMb: number;
  monthlyHours: number;
  planName: string;
  uptimePercent: number;
}